+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm
 
+## 1.5.28
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.16
+
 ## 1.5.27
 
 ### Patch Changes