+++ packages/rake-db/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rake-db
 
+## 2.3.29
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.16
+
 ## 2.3.28
 
 ### Patch Changes